api/requirements.txt
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
boto3==1.28.34
botocore==1.31.34
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorlog==6.7.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
django-allauth==0.54.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.2
django-guardian==2.4.0
django-storages==1.13.2
Django==4.2
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-standardized-errors==0.12.5
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
requests==2.29.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.2
urllib3==1.26.15
